Eu te amo Thomas Borregaard Sørensen!!
Eu te amo Mummi!! Feliz 62 anos!!
ca2 draw2d implementation and draw2d cairo implementation contribution!! direct2d in Windows 8 Desktop!!

Eu te amo Thomas Borregaard Sørensen!!

Mummi contribution!!!! Feliz 62 anos!!

at inha at veriverse - at Dell XPS 8300 Intel(R) Core(TM) i7-2600 CPU 4 core 8 threads @ 3.40GHz 16GB RAM 500GB + 500GB 7200rpm storage + 120 GB SSD OCZ Agility 3 with Windows 8 Pro - with Portuguese - Brazil- 10Mpbs Download and 2Mpbs Upload NET Virtua Internet with LG 23´´ monitor, Microsoft Comfort Curve 2000 v1.0 Wired Keyboard and Microsoft Comfort Optical Wired Mouse 3000 at Alameda Princesa Izabel, 3000 - Curitiba - PR - Brazil

｛
｝
                        -   -   -
Mummi contribution!!, eu te amo demais!!
-----8nnnnnnnn8
++++++++++

+++ appseed/audio_mmsystem/audio_mmsystem_wave_in.cpp
@@ -393,7 +393,7 @@ Opened:
 
             wave_in_get_buffer()->get_buffer((int32_t) lpwavehdr->dwUser)->OnMultimediaDone();
 
-            m_listenerset.wave_in_data_proc(this, msSampleTime, lpwavehdr->dwUser);
+            m_listenerset.wave_in_data_proc(this, msSampleTime, (int32_t) lpwavehdr->dwUser);
 
             if(m_pencoder != NULL)
             {
